+++ src/main/java/chess/controller/ChessSpringController.java
@@ -1,20 +1,10 @@
 package chess.controller;
 
 import chess.dto.ResultDto;
-import chess.dto.RoomInfoDto;
 import chess.dto.StatusDto;
 import chess.service.ChessGameService;
-import java.util.Arrays;
-import java.util.List;
-import javax.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
@@ -34,12 +24,6 @@ public class ChessSpringController {
         return modelAndView;
     }
 
-    @PostMapping("/")
-    public ResponseEntity<String> create(@Valid @RequestBody RoomInfoDto roomInfoDto) {
-        chessGameService.create(roomInfoDto.getTitle(), roomInfoDto.getPassword());
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
     @GetMapping("/board")
     public ModelAndView board(@RequestParam(name = "id") Long id) {
         ModelAndView modelAndView = new ModelAndView();
@@ -47,14 +31,6 @@ public class ChessSpringController {
                 .addObject("boardDto", chessGameService.getBoard(id));
         modelAndView.setViewName("index");
         return modelAndView;
-    }
-
-    @PatchMapping("/board/move")
-    public ResponseEntity<Boolean> move(@RequestParam(name = "id") Long id, @RequestBody String moveRequest) {
-        List<String> command = Arrays.asList(moveRequest.split(" "));
-        chessGameService.move(id, command.get(0), command.get(1));
-        final boolean gameEnd = chessGameService.isGameEnd(id);
-        return ResponseEntity.ok().body(gameEnd);
     }
 
     @GetMapping("/board/chess-status")
@@ -76,23 +52,5 @@ public class ChessSpringController {
                                 chessGameService.findWinner(id)));
         modelAndView.setViewName("result");
         return modelAndView;
-    }
-
-    @PostMapping("/board/end")
-    public ResponseEntity<String> end(@RequestParam(name = "id") Long id) {
-        chessGameService.end(id);
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    @PostMapping("/board/restart")
-    public ResponseEntity<String> restart(@RequestParam(name = "id") Long id) {
-        chessGameService.restart(id);
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    @DeleteMapping("/board")
-    public ResponseEntity<String> delete(@RequestParam(name = "id") Long id, @RequestBody String password) {
-        chessGameService.delete(id, password);
-        return ResponseEntity.status(HttpStatus.OK).build();
     }
 }
